% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_quantile.R
\name{tidy_quantile}
\alias{tidy_quantile}
\title{Tidy Quantile}
\usage{
tidy_quantile(data, var, length.out)
}
\arguments{
\item{data}{Dataframe/tibble which the variable is contained in}

\item{var}{What variable to make the quantiles from}

\item{length.out}{How many quantiles to create}
}
\value{
A quantile
}
\description{
This function takes in a variable from a tibble, and returns a quantile.
}
