% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_quantile.R
\name{tidy_quantile_st}
\alias{tidy_quantile_st}
\title{Tidy Quantile St}
\usage{
tidy_quantile_st(x, y, var, length.out)
}
\arguments{
\item{x}{The first sf Dataframe/tibble which the variable is contained in}

\item{y}{The second sf Dataframe/tibble which the variable is contained in}

\item{var}{What variable to make the quantiles from}

\item{length.out}{How many quantiles to create}
}
\value{
A quantile
}
\description{
This function loads into sf objects which are then merged to
create quantiles.
}
