% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_quantile.R
\name{tidy_tally_st}
\alias{tidy_tally_st}
\title{Tidy Talley st}
\usage{
tidy_tally_st(x, y, by)
}
\arguments{
\item{x}{The first sf Dataframe/tibble which the variable is contained in}

\item{y}{The second sf Dataframe/tibble which the variable is contained in}

\item{var}{What variable to make the quantiles from}

\item{length.out}{How many quantiles to create}
}
\value{
A quantile
}
\description{
This functcion returns a count for a given grouping variable and two sf objects that are
to be merged.
}
